[Chore] Split the audio and video halves out of the FFmpeg bridge
`media/ffi.rs` drops from 1 388 to 864 lines. Two more responsibilities
leave it, now that the audio encode path has tests that do not need a
fixture:

    audio  PCM encoding, resampling, the sample FIFO an encoder with a
           fixed frame size needs, and HCA-to-audio          492 lines
    video  scaling frames and choosing a pixel format         71 lines

The facade keeps the public conversions, the transcode drivers and the
codec plumbing both halves share. `audio` calls back into it for
`OutputCodec`, `ensure_ffmpeg_loaded` and `drain_encoder` -- written as
`use super::` with a comment, because encoding calls the drivers rather
than the reverse.

Direction is `error` (leaf) <- `avio`, `video` <- `raii` <- `audio`.

One thing worth recording: `cargo fix` removed the `ChannelLayout` import
because it is used only from the test module, which is the same shape as
the `PathBuf` loss this branch already had to repair. It is now imported
inside `mod tests`, where its users are.
